Add 'Restore Last Result' function & add to readme
For un-doing the 'Delete Last Result' button.

As written in the Readme: "In-case the last result was deleted in error (using the 'Delete Last Result' button as mentioned above), that result can actually be restored within the current session. Note that deleted results are not permanently stored, and will be lost if the app is restarted or if a new Result is taken."